rebase fork

+++ main/plugins/org.talend.designer.runprocess/src/main/java/org/talend/designer/runprocess/maven/listener/ProcessChangeListener.java
@@ -17,24 +17,28 @@ import java.beans.PropertyChangeListener;
 import java.util.List;
 import java.util.Set;
 
+import org.eclipse.core.resources.IContainer;
+import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IFolder;
 import org.eclipse.core.resources.IResource;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.eclipse.core.runtime.OperationCanceledException;
+import org.eclipse.core.runtime.Path;
 import org.eclipse.ltk.core.refactoring.resource.MoveResourceChange;
 import org.eclipse.ltk.core.refactoring.resource.RenameResourceChange;
 import org.talend.commons.exception.ExceptionHandler;
 import org.talend.core.GlobalServiceRegister;
 import org.talend.core.model.properties.Item;
-import org.talend.core.model.properties.ProcessItem;
 import org.talend.core.model.properties.Property;
 import org.talend.core.model.properties.RoutineItem;
 import org.talend.core.model.repository.ERepositoryObjectType;
 import org.talend.core.model.repository.IRepositoryViewObject;
 import org.talend.core.repository.model.ProxyRepositoryFactory;
+import org.talend.core.runtime.process.TalendProcessOptionConstants;
 import org.talend.core.ui.ITestContainerProviderService;
+import org.talend.designer.maven.model.TalendMavenConstants;
 import org.talend.designer.maven.tools.AggregatorPomsHelper;
 import org.talend.designer.maven.tools.BuildCacheManager;
 import org.talend.designer.runprocess.java.TalendJavaProjectManager;
@@ -48,7 +52,7 @@ public class ProcessChangeListener implements PropertyChangeListener {
 
     private AggregatorPomsHelper helper;
 
-    private List<ERepositoryObjectType> allProcessTypes = ERepositoryObjectType.getAllTypesOfProcess();
+    private List<ERepositoryObjectType> allProcessTypes;
 
     @Override
     public void propertyChange(PropertyChangeEvent event) {
@@ -65,11 +69,14 @@ public class ProcessChangeListener implements PropertyChangeListener {
                     caseMove(oldValue, newValue);
                 } else if (propertyName.equals(ERepositoryActionName.DELETE_FOREVER.getName())) {
                     caseDeleteForever(newValue);
-                } else if (propertyName.equals(ERepositoryActionName.FOLDER_RENAME.getName())) {
+                } else if (propertyName.equals(ERepositoryActionName.FOLDER_RENAME.getName())
+                        || propertyName.equals(ERepositoryActionName.JOBLET_FOLDER_RENAME.getName())) {
                     caseFolderRename(oldValue, newValue);
-                } else if (propertyName.equals(ERepositoryActionName.FOLDER_MOVE.getName())) {
+                } else if (propertyName.equals(ERepositoryActionName.FOLDER_MOVE.getName())
+                        || propertyName.equals(ERepositoryActionName.JOBLET_FOLDER_MOVE.getName())) {
                     caseFolderMove(oldValue, newValue);
-                } else if (propertyName.equals(ERepositoryActionName.FOLDER_DELETE.getName())) {
+                } else if (propertyName.equals(ERepositoryActionName.FOLDER_DELETE.getName())
+                        || propertyName.equals(ERepositoryActionName.JOBLET_FOLDER_DELETE.getName())) {
                     caseFolderDelete(oldValue, newValue);
                 } else if (propertyName.equals(ERepositoryActionName.SAVE.getName())
                         || propertyName.equals(ERepositoryActionName.CREATE.getName())) {
@@ -93,32 +100,26 @@ public class ProcessChangeListener implements PropertyChangeListener {
         // oldValue: [oldName, oldVersion], newValue: property
         if (oldValue instanceof String[] && newValue instanceof Property) {
             Property property = (Property) newValue;
-            if (property.getItem() instanceof ProcessItem) {
-                ITestContainerProviderService service = getTestContainerProviderService();
-                boolean isTestCase = false;
-                if (service != null) {
-                    isTestCase = service.isTestContainerItem(property.getItem());
+            ITestContainerProviderService service = getTestContainerProviderService();
+            if (service != null && service.isTestContainerItem(property.getItem())) {
+                // FIXME
+            } else if (isNeedUpdateItem(property.getItem())) {
+                String[] oldFields = (String[]) oldValue;
+                if (oldFields.length == 0) {
+                    // Do nothing case import project
+                    return;
                 }
-                if (isTestCase) {
-                    // FIXME
-                } else {
-                    String[] oldFields = (String[]) oldValue;
-                    if (oldFields.length == 0) {
-                        // Do nothing case import project
-                        return;
-                    }
-                    String oldName = oldFields[0];
-                    String oldVersion = oldFields[1];
-                    if (!oldName.equals(property.getLabel())) {
-                        // job name change, will change all old version job name
-                        // delete all version old job projects physically(need to recreate pom), and create new for all.
-                        TalendJavaProjectManager.deleteAllVersionTalendJobProject(property.getId(), oldName, true);
-                        TalendJavaProjectManager.createAllVersionTalendJobProject(property.getId());
-                    } else if (!oldVersion.equals(property.getVersion())) {
-                        // version change, will create new item
-                        // create new job project.
-                        TalendJavaProjectManager.generatePom((ProcessItem) property.getItem());
-                    }
+                String oldName = oldFields[0];
+                String oldVersion = oldFields[1];
+                if (!oldName.equals(property.getLabel())) {
+                    // job name change, will change all old version job name
+                    // delete all version old job projects physically(need to recreate pom), and create new for all.
+                    TalendJavaProjectManager.deleteAllVersionTalendJobProject(property.getId(), oldName, true);
+                    TalendJavaProjectManager.createAllVersionTalendJobProject(property.getId());
+                } else if (!oldVersion.equals(property.getVersion())) {
+                    // version change, will create new item
+                    // create new job project.
+                    TalendJavaProjectManager.generatePom(property.getItem());
                 }
             }
         }
@@ -129,7 +130,7 @@ public class ProcessChangeListener implements PropertyChangeListener {
         if (oldValue instanceof IRepositoryViewObject && newValue instanceof IPath[]) {
             // delete all version job project, create all version new projects
             IRepositoryViewObject obj = (IRepositoryViewObject) oldValue;
-            if (obj.getProperty().getItem() instanceof ProcessItem) {
+            if (isNeedUpdateItem(obj.getProperty().getItem())) {
                 // TalendJavaProjectManager.deleteAllVersionTalendJobProject(obj.getId(), null, false);
                 IPath sourcePath = ((IPath[]) newValue)[0];
                 IPath targetPath = ((IPath[]) newValue)[1];
@@ -145,8 +146,13 @@ public class ProcessChangeListener implements PropertyChangeListener {
                         String jobProjectFolderName = AggregatorPomsHelper.getJobProjectFolderName(property);
                         IFolder sourceFolder = processTypeFolder.getFolder(sourcePath).getFolder(jobProjectFolderName);
                         if (sourceFolder.exists()) {
+                            sourceFolder.getProject().refreshLocal(IResource.DEPTH_INFINITE, null);
+                            AggregatorPomsHelper
+                                    .removeFromParentModules(sourceFolder.getFile(TalendMavenConstants.POM_FILE_NAME));
                             MoveResourceChange change = new MoveResourceChange(sourceFolder, targetFolder);
                             change.perform(new NullProgressMonitor());
+                            AggregatorPomsHelper.addToParentModules(
+                                    targetFolder.getFolder(sourceFolder.getName()).getFile(TalendMavenConstants.POM_FILE_NAME));
                         }
                     }
                 } catch (Exception e) {
@@ -160,15 +166,11 @@ public class ProcessChangeListener implements PropertyChangeListener {
         // newValue: obj
         if (newValue instanceof IRepositoryViewObject) {
             Property property = ((IRepositoryViewObject) newValue).getProperty();
-            if (property != null) {
+            if (property != null && property.getItem() != null) {
                 ITestContainerProviderService service = getTestContainerProviderService();
-                boolean isTestCase = false;
-                if (service != null && property.getItem() != null) {
-                    isTestCase = service.isTestContainerItem(property.getItem());
-                }
-                if (isTestCase) {
+                if (service != null && service.isTestContainerItem(property.getItem())) {
                     // FIXME testcase should regenerate pom to remove extra dependencies.
-                } else {
+                } else if (isNeedUpdateItem(property.getItem())) {
                     // delete all version old job projects physically, won't check to remove parent folder
                     TalendJavaProjectManager.deleteAllVersionTalendJobProject(property.getId(), null, true);
                 }
@@ -184,16 +186,67 @@ public class ProcessChangeListener implements PropertyChangeListener {
             Object[] objects = (Object[]) newValue;
             String newName = (String) objects[0];
             ERepositoryObjectType processType = (ERepositoryObjectType) objects[1];
-            if (allProcessTypes.contains(processType)) {
-                // TalendJavaProjectManager.deleteTalendJobProjectsUnderFolder(processType, folderPath, false);
+            if (getAllProcessTypes().contains(processType)) {
                 IFolder sourceFolder = getAggregatorPomsHelper().getProcessFolder(processType).getFolder(folderPath);
                 NullProgressMonitor monitor = new NullProgressMonitor();
                 try {
+                    IContainer parent = sourceFolder.getParent();
                     sourceFolder.refreshLocal(IResource.DEPTH_INFINITE, monitor);
+                    removeFromParentSourceFolder(sourceFolder);
                     RenameResourceChange change = new RenameResourceChange(sourceFolder.getFullPath(), newName);
                     change.perform(monitor);
+                    IFolder newFolder = parent.getFolder(new Path(newName));
+                    addToParentInNewFolder(newFolder);
                 } catch (CoreException e) {
                     ExceptionHandler.process(e);
+                }
+            }
+        }
+    }
+
+    /**
+     * DOC nrousseau Comment method "addToParentInNewFolder".
+     * 
+     * @param newFolder
+     * @throws CoreException
+     */
+    private void addToParentInNewFolder(IFolder newFolder) throws CoreException {
+        for (IResource res : newFolder.members()) {
+            if (res instanceof IFolder) {
+                IFolder currentFolder = (IFolder) res;
+                IFile pomFile = currentFolder.getFile(TalendMavenConstants.POM_FILE_NAME);
+                if (pomFile.exists()) {
+                    try {
+                        AggregatorPomsHelper.addToParentModules(pomFile);
+                    } catch (Exception e) {
+                        ExceptionHandler.process(e);
+                    }
+                } else {
+                    addToParentInNewFolder(currentFolder);
+                }
+            }
+        }
+    }
+
+    /**
+     * DOC nrousseau Comment method "removeFromParentSourceFolder".
+     * 
+     * @param sourceFolder
+     * @throws CoreException
+     */
+    private void removeFromParentSourceFolder(IFolder sourceFolder) throws CoreException {
+        for (IResource res : sourceFolder.members()) {
+            if (res instanceof IFolder) {
+                IFolder currentFolder = (IFolder) res;
+                IFile pomFile = currentFolder.getFile(TalendMavenConstants.POM_FILE_NAME);
+                if (pomFile.exists()) {
+                    try {
+                        AggregatorPomsHelper.removeFromParentModules(pomFile);
+                    } catch (Exception e) {
+                        ExceptionHandler.process(e);
+                    }
+                } else {
+                    removeFromParentSourceFolder(currentFolder);
                 }
             }
         }
@@ -206,15 +259,18 @@ public class ProcessChangeListener implements PropertyChangeListener {
             IPath sourcePath = ((IPath[]) oldValue)[0];
             IPath targetPath = ((IPath[]) oldValue)[1];
             ERepositoryObjectType processType = (ERepositoryObjectType) newValue;
-            // TalendJavaProjectManager.deleteTalendJobProjectsUnderFolder(processType, sourcePath, false);
-            IFolder processTypeFolder = getAggregatorPomsHelper().getProcessFolder(processType);
-            IFolder sourceFolder = processTypeFolder.getFolder(sourcePath);
-            IFolder targetFolder = processTypeFolder.getFolder(targetPath);
-            MoveResourceChange change = new MoveResourceChange(sourceFolder, targetFolder);
-            try {
-                change.perform(new NullProgressMonitor());
-            } catch (OperationCanceledException | CoreException e) {
-                ExceptionHandler.process(e);
+            if (getAllProcessTypes().contains(processType)) {
+                IFolder processTypeFolder = getAggregatorPomsHelper().getProcessFolder(processType);
+                IFolder sourceFolder = processTypeFolder.getFolder(sourcePath);
+                try {
+                    removeFromParentSourceFolder(sourceFolder);
+                    IFolder targetFolder = processTypeFolder.getFolder(targetPath);
+                    MoveResourceChange change = new MoveResourceChange(sourceFolder, targetFolder);
+                    change.perform(new NullProgressMonitor());
+                    addToParentInNewFolder(targetFolder.getFolder(sourceFolder.getName()));
+                } catch (OperationCanceledException | CoreException e) {
+                    ExceptionHandler.process(e);
+                }
             }
         }
     }
@@ -225,15 +281,17 @@ public class ProcessChangeListener implements PropertyChangeListener {
             // delete all version job projects under it physically, remove this folder.
             IPath folderPath = (IPath) oldValue;
             ERepositoryObjectType processType = (ERepositoryObjectType) newValue;
-            TalendJavaProjectManager.deleteTalendJobProjectsUnderFolder(processType, folderPath, true);
+            if (getAllProcessTypes().contains(processType)) {
+                TalendJavaProjectManager.deleteTalendJobProjectsUnderFolder(processType, folderPath, true);
+            }
         }
     }
 
     private void caseCreateAndSave(String propertyName, Object newValue) {
         if (newValue instanceof Item) {
-            if (newValue instanceof ProcessItem) {
+            if (isNeedUpdateItem((Item) newValue)) {
                 if (propertyName.equals(ERepositoryActionName.SAVE.getName())) {
-                    TalendJavaProjectManager.generatePom((ProcessItem) newValue);
+                    TalendJavaProjectManager.generatePom((Item) newValue);
                 }
             } else if (newValue instanceof RoutineItem) {
                 updateCodesChange((RoutineItem) newValue);
@@ -246,8 +304,8 @@ public class ProcessChangeListener implements PropertyChangeListener {
         if (newValue instanceof Set) {
             Set<Item> importItems = (Set<Item>) newValue;
             for (Item item : importItems) {
-                if (item instanceof ProcessItem) {
-                    TalendJavaProjectManager.generatePom((ProcessItem) item);
+                if (isNeedUpdateItem(item)) {
+                    TalendJavaProjectManager.generatePom(item, TalendProcessOptionConstants.GENERATE_NO_CODEGEN);
                 } else if (item instanceof RoutineItem) {
                     updateCodesChange((RoutineItem) item);
                 }
@@ -258,6 +316,14 @@ public class ProcessChangeListener implements PropertyChangeListener {
     private void updateCodesChange(RoutineItem codesItem) {
         ERepositoryObjectType type = ERepositoryObjectType.getItemType(codesItem);
         BuildCacheManager.getInstance().updateCodesLastChangeDate(type, codesItem.getProperty());
+    }
+
+    private boolean isNeedUpdateItem(Item item) {
+        ERepositoryObjectType type = ERepositoryObjectType.getItemType(item);
+        if (type != null) {
+            return getAllProcessTypes().contains(type);
+        }
+        return false;
     }
 
     private ITestContainerProviderService getTestContainerProviderService() {
@@ -273,6 +339,13 @@ public class ProcessChangeListener implements PropertyChangeListener {
             helper = new AggregatorPomsHelper();
         }
         return helper;
+    }
+
+    private List<ERepositoryObjectType> getAllProcessTypes() {
+        if (allProcessTypes == null) {
+            allProcessTypes = ERepositoryObjectType.getAllTypesOfProcess2();
+        }
+        return allProcessTypes;
     }
 
 }
